Support root CA variable for OCSP validation
- veritasos/ingress-nginx forked from kubernetes/ingress-nginx
- develop branch created off commit ce418168f4f20a16a8622f17f756b578c1418d29
- committing on behalf of Kushal Kumaran diff from these commits (includes original commit messages):
1) https://stash.veritas.com/users/kushal.kumaran/repos/ingress-nginx/commits/aba1025cc90b1f85e959d895459cbcb90fb1cc2d

set nginx variable with path to CA root

OCSP validation code needs to be able to read the issuer certificate
so that it can be included in the OCSP request.  This change sets a
variable that the lua code can then access to get the path where the
root certificate is stored.

2) https://stash.veritas.com/users/kushal.kumaran/repos/ingress-nginx/commits/9764311b1e921cedba3fa48b0130cab954108fcb

add lua-resty-url

The lua-resty-url library provides functions for parsing URIs into its
components that can be passed into HTTP client library.

3) https://stash.veritas.com/users/kushal.kumaran/repos/ingress-nginx/commits/4de0bc08e8bf3e27c64a75d0a01448e1ff7c6f55

add lua-resty-http

The lua-resty-http module is a lua HTTP client.  This library allows
lua code running in nginx configuration to make HTTP calls, such as to
OCSP servers.