allow nginx setup with FIPS-validated OpenSSL library
This patch allows the ingress container to be built with a
FIPS-validated OpenSSL library.  The FIPS_OPENSSL environment variable
must be set when building the container.  The library that the OS
provides is replaced with the OpenSSL from the provided URL.  libcurl,
which depends on OpenSSL is rebuilt against the replaced version as
well.